Better errors, added challengeId to params
