+++ src/routes/(ctf_platform)/ctf/[ctf_id=ctf_id]/organizer/approve/[challengeId=challengeId]/+page.svelte
@@ -1,11 +1,20 @@
 <script lang="ts">
 	import { enhance } from '$app/forms';
-	import { categories } from '$lib/db/constants';
 	let { data, form } = $props();
 	let { translations, unapprovedChallenge } = data;
 
 	type resource_type = 'File' | 'Command' | 'Website';
 
+	let categories = [
+		'crypto',
+		'forensics',
+		'introduction',
+		'misc',
+		'osint',
+		'pwn',
+		'reversing',
+		'web'
+	];
 	let mainCategory: string = $state(categories[3]);
 	if (unapprovedChallenge?.challenge_category) {
 		mainCategory = unapprovedChallenge?.challenge_category;
