+++ src/routes/(wargame_platform)/admin/approve/[challengeId=challengeId]/+page.server.ts
@@ -12,29 +12,19 @@ import { categories } from '$lib/db/constants';
 
 export const load: PageServerLoad = async ({ locals, params }) => {
 	const user = locals.user;
-	const ctfId = Number(params.ctf_id);
 	const challengeId = params.challengeId;
 
 	if (!user) {
 		return redirect(303, '/login');
 	}
 
-	const org = await db
-		.selectFrom('ctf_organizers')
-		.where('ctf', '=', ctfId)
-		.where('user_id', '=', user.id)
-		.executeTakeFirst();
-
-	const isOrg = org !== undefined;
-
-	if (!isOrg && !user.is_admin) {
-		return error(401, 'User not organizer for this CTF or admin');
+	if (!user.is_admin) {
+		return error(401, 'User not admin');
 	}
 
 	const unapprovedChallenge = await db
 		.selectFrom('challenges as ch')
 		.where('challenge_id', '=', challengeId)
-		.where('ch.ctf', '=', ctfId)
 		.where('ch.approved', '=', false)
 		.leftJoin('flag as f', 'ch.flag', 'f.id')
 		.leftJoin('users as a', 'ch.author', 'a.id')
@@ -78,6 +68,9 @@ export const load: PageServerLoad = async ({ locals, params }) => {
 		])
 		// .orderBy('ch.points', 'asc')
 		.executeTakeFirst();
+	if (unapprovedChallenge === undefined) {
+		return error(404, { message: 'Challenge not found' });
+	}
 
 	return { unapprovedChallenge };
 };
@@ -86,23 +79,14 @@ export const actions = {
 	default: async ({ request, params, locals }) => {
 		try {
 			const user = locals.user;
-			const ctfId = Number(params.ctf_id);
 			const challengeId = params.challengeId;
 
 			if (!user) {
 				return redirect(304, '/login');
 			}
 
-			const org = await db
-				.selectFrom('ctf_organizers')
-				.where('ctf', '=', ctfId)
-				.where('user_id', '=', user.id)
-				.executeTakeFirst();
-
-			const isOrg = org !== undefined;
-
-			if (!isOrg && !user.is_admin) {
-				return fail(401, { message: 'User not organizer for this CTF or admin' });
+			if (!user.is_admin) {
+				return error(401, 'User not admin');
 			}
 
 			const formData = await request.formData();
@@ -120,7 +104,7 @@ export const actions = {
 
 			const points = Number(formData.get('points'));
 			if (points < 0) {
-				return fail(422, { message: 'Points must be a non-negative integer' });
+				return fail(404, { message: 'Points must be a non-negative integer' });
 			}
 
 			const mainCategory: Category = validateCategory(
@@ -152,7 +136,6 @@ export const actions = {
 					approved: true
 				})
 				.where('challenge_id', '=', challengeId)
-				.where('ctf', '=', ctfId)
 				.returning('flag')
 				.executeTakeFirst();
 
@@ -298,7 +281,6 @@ export const actions = {
 
 			return { success: true, message: 'Challenge successfully approved' };
 		} catch (err) {
-			console.error(err);
 			return fail(500, { message: err.message });
 		}
 	}
