+++ src/routes/(wargame_platform)/admin/approve/[challengeId=challengeId]/+page.svelte
